refactor(signaler): move signaler to observation
