+++ packages/runtime/src/binding/ast.ts
@@ -2,11 +2,11 @@ import { IIndexable, IServiceLocator, PLATFORM, Reporter, StrictPrimitive } from
 import { IBindScope } from '../lifecycle';
 import { Collection, IBindingContext, IOverrideContext, IScope, LifecycleFlags, ObservedCollection } from '../observation';
 import { BindingContext } from '../observation/binding-context';
+import { ISignaler } from '../observation/signaler';
 import { BindingBehaviorResource } from '../resources/binding-behavior';
 import { IValueConverter, ValueConverterResource } from '../resources/value-converter';
 import { IBinding } from './binding';
 import { IConnectableBinding } from './connectable';
-import { ISignaler } from './signaler';
 
 export type IsPrimary = AccessThis | AccessScope | ArrayLiteral | ObjectLiteral | PrimitiveLiteral | Template;
 export type IsLiteral = ArrayLiteral | ObjectLiteral | PrimitiveLiteral | Template;
